[Refactor] : api -> apis 로 폴더이름 수정
폴더이름을 apis로 수정하고 절대경로를 이용했습니다.
axios create를 이용한 공통 인스턴스를 이용하도록 api코드들을 수정했습니다.
